+++ CHANGELOG.md
@@ -17,6 +17,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Added `compas_viewer.components.SceneForm`.
 * Added `callback` for `compas_viewer.components.TreeForm`.
 * Added click-boxes of `show` and `locked` for `compas_viewer.components.SceneForm`.
+* Added observer pattern to of `show` and `locked` for `compas_viewer.Scene.scene`.
+* Added Time Debounce pattern to of `show` and `locked` for `compas_viewer.Scene.scene`.
 
 ### Changed